refactor(common): move `ibis.collections.DisjointSet` to `ibis.common.egraph`
